refactoring sonarqube vunnarabilities/bugs and code smells

+++ app/src/main/java/code/name/monkey/retromusic/util/CalendarUtil.java
@@ -19,7 +19,7 @@ import java.util.GregorianCalendar;
 
 /** @author Eugene Cheung (arkon) */
 public class CalendarUtil {
-  private static final long MS_PER_MINUTE = 60 * 1000;
+  private static final long MS_PER_MINUTE = (long)60 * 1000;
   private static final long MS_PER_DAY = 24 * 60 * MS_PER_MINUTE;
 
   private Calendar calendar;
